// app/cdk/types/types.ts
export type SleeperPlayer = {
    player_id: string;
    full_name: string;
    first_name: string;
    last_name: string;
    search_full_name: string;
    search_first_name: string;
    search_last_name: string;
    age: number;
    birth_date: string | null;
    birth_city: string | null;
    birth_state: string | null;
    birth_country: string | null;
    college: string | null;
    high_school: string | null;
    height: string | null;
    weight: string | null;
    position: string;
    fantasy_positions: string[];
    team: string | null;
    team_abbr: string | null;
    number: number | null;
    status: string;
    years_exp: number;
    active: boolean;
    injury_status: string | null;
    injury_notes: string | null;
    injury_start_date: string | null;
    injury_body_part: string | null;
    practice_description: string | null;
    practice_participation: string | null;
    news_updated: string | null;
    team_changed_at: string | null;
    depth_chart_order: number | null;
    depth_chart_position: string | null;
    competitions: string[];
    sport: string;
    espn_id: number | null;
    yahoo_id: number | null;
    rotoworld_id: number | null;
    rotowire_id: number | null;
    fantasy_data_id: number | null;
    stats_id: number | null;
    gsis_id: string | null;
    sportradar_id: string | null;
    pandascore_id: string | null;
    opta_id: string | null;
    swish_id: string | null;
    oddsjam_id: string | null;
    metadata?: {
        channel_id?: string;
        [key: string]: unknown;
    };

    hashtag: string;
    search_rank: number;
};